fixed many warning

+++ dream/src/GUI-QT/fdrmdialog.cpp
@@ -493,7 +493,7 @@ void FDRMDialog::showTextMessage(const QString& textMessage)
     TextTextMessage->setEnabled(TRUE);
 
     QString formattedMessage = "";
-    for (size_t i = 0; int(i) < textMessage.length(); i++)
+    for (int i = 0; i < (int)textMessage.length(); i++)
     {
         switch (textMessage.at(i).unicode())
         {
